Add docs to AudioHelper
