+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/AudioHelper.kt
@@ -25,20 +25,32 @@ import android.media.AudioTrack
 import android.media.MediaRecorder
 import android.media.audiofx.AcousticEchoCanceler
 import androidx.annotation.RequiresPermission
-import kotlinx.coroutines.delay
+import com.google.firebase.vertexai.common.util.readAsFlow
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-import kotlinx.coroutines.flow.flow
 
+/**
+ * Helper class for recording audio and playing back a separate audio track at the same time.
+ *
+ * @see AudioHelper.build
+ */
 @PublicPreviewAPI
 internal class AudioHelper(
-  // Record for recording the user's mic
+  /** Record for recording the System microphone. */
   private val recorder: AudioRecord,
-  // Track for playing back what the model says
+  /** Track for playing back what the model says. */
   private val playbackTrack: AudioTrack,
 ) {
   private var released: Boolean = false
 
+  /**
+   * Release the system resources on the recorder and playback track.
+   *
+   * Once an [AudioHelper] has been "released", it can _not_ be used again.
+   *
+   * This method can safely be called multiple times, as it won't do anything if this instance has
+   * already been released.
+   */
   fun release() {
     if (released) return
     released = true
@@ -47,24 +59,48 @@ internal class AudioHelper(
     playbackTrack.release()
   }
 
+  /**
+   * Play the provided audio data on the playback track.
+   *
+   * Does nothing if this [AudioHelper] has been [released][release].
+   */
   fun playAudio(data: ByteArray) {
     if (released) return
 
     playbackTrack.write(data, 0, data.size)
   }
 
+  /**
+   * Pause the recording of the microphone, if it's recording.
+   *
+   * Does nothing if this [AudioHelper] has been [released][release].
+   *
+   * @see resumeRecording
+   */
   fun pauseRecording() {
     if (released || recorder.recordingState == AudioRecord.RECORDSTATE_STOPPED) return
 
     recorder.stop()
   }
 
+  /**
+   * Resumes the recording of the microphone, if it's not already running.
+   *
+   * Does nothing if this [AudioHelper] has been [released][release].
+   *
+   * @see pauseRecording
+   */
   fun resumeRecording() {
     if (released || recorder.recordingState == AudioRecord.RECORDSTATE_RECORDING) return
 
     recorder.startRecording()
   }
 
+  /**
+   * Start perpetually recording the system microphone, and return the bytes read in a flow.
+   *
+   * Returns an empty flow if this [AudioHelper] has been [released][release].
+   */
   fun listenToRecording(): Flow<ByteArray> {
     if (released) return emptyFlow()
 
@@ -74,8 +110,17 @@ internal class AudioHelper(
   }
 
   companion object {
+    /**
+     * Creates an instance of [AudioHelper] with the track and record initialized.
+     *
+     * A separate build method is necessary so that we can properly propagate the required manifest
+     * permission, and throw exceptions when needed.
+     *
+     * It also makes it easier to read, since the long initialization is separate from the
+     * constructor.
+     */
     @RequiresPermission(Manifest.permission.RECORD_AUDIO)
-    fun Build(): AudioHelper {
+    fun build(): AudioHelper {
       val playbackTrack =
         AudioTrack(
           AudioAttributes.Builder().setUsage(AudioAttributes.USAGE_VOICE_COMMUNICATION).build(),
@@ -125,27 +170,6 @@ internal class AudioHelper(
       }
 
       return AudioHelper(recorder, playbackTrack)
-    }
-  }
-}
-
-internal val AudioRecord.minBufferSize: Int
-  get() = AudioRecord.getMinBufferSize(sampleRate, channelConfiguration, audioFormat)
-
-internal fun AudioRecord.readAsFlow() = flow {
-  val buffer = ByteArray(minBufferSize)
-
-  while (true) {
-    if (recordingState != AudioRecord.RECORDSTATE_RECORDING) {
-      delay(1)
-      continue
-    }
-
-    val bytesRead = read(buffer, 0, buffer.size)
-    if (bytesRead > 0) {
-      emit(buffer.copyOf(bytesRead))
-    } else {
-      delay(1)
     }
   }
 }
